flask_shit/requirements.txt
mediapipe
flask-socketio